Spreadsheet communication update
I added a script that uses the Google Sheets Api to access all the data on a sheet at one time.
It then takes that data and breaks it down into the rows, which then become arrays to hold all the data for that particular object.
It then hands the data over in a List of all these arrays.
From there I have called the games main script to take the data and check whether there is a playerprefs value associated with it, and if not, it creates one for it.
If there was a value there, it instead sets the games state to these values.
I might want to look into a json file to hold the offline information rather than using playerprefs.