Scene Building
Started to create a main scene area, where I can create an amount of UI elements on the top for whatever we want it to be.
Realized I also needed to create a data resource for each button so they can know what to populate.
In said script, I started to give assignment to resources, and break down their respective requirements so that we know how to build them in the future.