Resource Update
Added in a UI button prefab that will be the main focal point of a "build", also added in a button prefab of each "dependent" of the build.
In the Resource and ResourceData scripts I added in ways to adjust their respective information as well as how to find all the dependence of the major resource on screen.
Upon finding them, it will instantiate all dependency from smallest to the greatest in a clockwise fashion giving each their respective data type.
I have yet to actually test this, so after this I will test it.