Buttons, Main UI Display and SaveSystem Update
For each of the button prefabs(only two), I made sure that when you hover over them, they show a panel with details of the item type and dependencies.
The all rotate around the appropriate amount and stay the right direction.
Adding the panel and hoverAble scripts. Each of the resource aspects needed some improvement as well.
The main script had an issue with saving an empty file when exiting, so I fixed it to save everything before exit.