SheetReader/Resource/Main Update
I changed the sheet ID in sheetreader as well as the row to start pulling content.
In ResourceInfoPanel I added a legit call to real data for the descriptions.
I changed autoTime to CraftTime, and had to adjust it in ResourcedisplayInfo.
ResourceData got a slot for descriptions, groups, consumable and nonconsumable requirements, itemsToBeGained, commandsOnPressed, commandsOnCreated, image name, sound name and acheivement call.
Resource now looks for nonconsumables as dependencies, and adds them into the new list.
I also added a call to become visible which tells Main to start a coroutine of that resource.
Main got a lot of changes, a string array containing names and an associated int array to hold a qued amount of resources.
I also added a DoNotdestroyOnLoadCall because I don't think that the game should keep rebuilding info after the game has started.
I also added a Delete data button when pressing the '1' key.
Main has a start Que Update, QueUpdate, AddToQue call.
I also had to update the building of assets in main, so that it can now appropriately see all assets as they are in the sheet.
A quality of life change was making a helper function called CreateResourcePanelReferenceAndSave which does exactly that.