SaveSystem/Resource/ResourcePanel Update
With the resource info panel prefab, I added a text for how much you have of the current resource that you are looking at.
I also needed to add that field into the InfoPanel script and update it with the other values when necessary.
In Main, I added a couple debug logs to better see what all is happening.
Also modified how load looks at the bool for visibility, it was always coming out false. Now it reads correctly, and also correctly loads previous amounts.
In resource I pulled the check for consumable and nonconsumable dependencies out from the same loop. Each has their own and works much better.
Also added in a check for groups on dependency, so that tools will not be consumed as they are non-consumable,
Also changed the button rotation check for stopping at the actual last one, rather than the second to last one. Will still need to do something if there are more than 8 dependencies.
I created an IEnumerator for running an update once a change to the numbers are supposed to be called so that it waits for the amount of time then lets you know something has changed.