Button and Debug Update
I gave the BuildableResource prefab the dependencyprefab.
Created a message board prefab which could be called and created at any time, and can change the title/type and description text at any time.
I gave the Dependence prefab itself to create more for the panels. (May want to call this when asked to rather than creating all as soon as it is created the first time)
Resource info panel, and hoverable needed some fenagling to get the loop to create all that is needed.
Added a Debug section to the code in main to adjust any resource on the fly.
Added a dependency check that any script can do as long as it can access main and can call for those dependencies and their amounts.
Set up a "drop" call that takes the string and grabs all resources and adjusts them according to what was given.
Resource also needed some weird hoops to get the button/panel loop correct.
Added gameElementType and atMostAmount to Data(needed to add that part to the deserialization part for saving and loading later).
Set up a Weighted system.