Save and Load Update
I created a try catch call for whether the game was able to connect to the internet to build the resources or not.
It now can build the game from memory should the game be offline.
The save string had an extra ; at the end and created an empty object when loading, but that is now fixed as well. Game runs perfectly fine online as well as offline.
Also added a force recreate from sheet data and a manual delete file to help speed up any process we might need to reset everything.
Also realized I spelled Description wrong and fixed it in the ResourceData and all placed that reference it.