ResourceData Update
I added a new string value to hold onto in ResourceData called buildables.
In main, it will grab all resources and create a string for their respective builds. So that later we can utilize this list and populate a field to show both what is required to build the resource as well as what the resource can build into.