ResourceData and Main Update
I added a call to set all variables that each ResourceData holds so that I can change them when I want to.
A use case for this is when we are building all resources and I take what is stored in memory, but I have a changed value on the database, I will grab that resource and check each values until I find something to change and will make it the same as the cloud.
I also checked for an out of index range for when the values are not populated for the change group, so it will skip the error and move on down the code.
Added in a CreateAllBuildables string for each resource.
Also checking if either the a consumable or nonconsumable requirement was changed, if so, all buildable strings could be affected and need updated as well.
I also added in regions to Main because it is starting to get very large.