OverWorld Update
Created a galaxy and supercluster prefab to start experimenting with seeded generation.
Changed a few things in the Main scene.
Created a new script for the DepthInteractions, basically telling Main which object of the space objects it was.
SaveSystem has been given a new things to save, could easily save it in the same file as the resource library, but keeping it separate for now.
Main got most the update, it got several changes to the load and save methods.
It also got several new methods to build the generation of the universe, based on the position of each object in their own array then broken down as an integer from the HashCode of the string of all the locations.
Currently I have been able to go down the randomization into deeper parts and get the same thing as well as going backwards to the previous parent level based on the string input.
I was having some trouble which will probably be easily fixed with the highest level though.