Planetary Creation and space Traversal update
Added in all the images for each of the space elements.
Added the unclicked, clickable and conquered images for the hex tile.
I created a material for each of the images to attach them to each of the quads.
Had to make a few changes to the main scene so that it will hold all the info needed for the new stuff added.
Added a CameraController script that alters and allows control of the camera movement on planets.
Created a HexTileInfo script, it currently just swaps the material on the hex whether the mouse is hovering over it or not.
Main script had a huge update in script since it now creates planets, it currently creates a nice diamond pattern of 100 hex tiles.
Yet to create different data for the map, as well as saving and loading that  data.