SaveSystem and Planet Building Update
I changed locations of the GoBack and To the top button locations so they are less likely to interact with the background ui.
In CameraController I have it now listed for a target, it will focus on the target and change the camera orthographic scale to 1.5 while in a map.
In the interactable script it no longer has a hovering bool.
In the hextileInfo script, it now knows if it is next to a conquered spot, and if so will update itself to be an interactable location.
I added in a setting to starting point call and a force state from reload call.
It also right now just conquers when clicked and is clickable and sends a message to the others to see whether they should update as well.
In savesystem I broke up all the info into FileData, Address, Resources, and Tiles.
When saving each will be assigned based on what is supposed to be saved, then will combine them in a manner that is conducive to the file.
There will be a ResourceLibraryFile, a current AddressFile, an AddressBookFile for all locations we have actually been to, and a location file for all the data on a given location, which is named on it's location information.
SheetReader had an unnecessary (Stream)cast so I removed it.
Main got the overhaul again:
Added a HexTileInfo[], ResourceData[] called LocationResources, and a string list for an address book.
I rebuilt the first call of the game, it either builds/updates the resourcelibrary if the sheet was accessed, otherwise it recalls the file. Currently don't have a fail safe, should it try to grab the save and not be connected and no save is there.
Then we have a building of a whole new resource library but is location specific and will get rebuilt upon going back to the location.
Also had to change the saving when the application quit, as well as when it should be saving info.