Major System Update
Changed a few overhead things in the Main scene.
CameraController can now zoom in and out in all situations.
It can also remember it's last state when playing and load back to the position.
DepthInteraction can now manually or automatically deactivate making it so you can't just click on something and make the animation play over and over or break.
Tile script can now further remember what type of tile it was and adjust itself accordingly when told.
Created a LocationManager that is the brain of a location and will hold all information about it. It also builds it's own level now instead of Main.
It can also save all the states and rebuild from memory of a location.
It also can turn itself on and off when switching between planets.
Main was updated to no longer control level building but still acts as a port for all information to pass through.
It also takes all information given and sends it to the saving system.
Savesystem needed a new function to save camera information.
Removed unnecessary material from the tile material list.
Added sprites for tiles.