Map Update
Removed Materials until we had two.
Swapped out the settings on all the textures.
In the tile prefab it holds two materials associated with the two material pieces they use.
It also holds all the textures now and adjusts them as needed.
The hextile now holds all textures and materials it needs, it also knows it's neighbors and can save them and load them.
Location brain now holds all tilelocations in a vector2 array. Upon building all new tiles it will find the neighbors and assign them.
When loading, it will send the string for that particular tile to it after creating it.
Also put a bounds on the camera zoom so it can't go backwards or zoom out too far.