Planetary Update
We now have a landing animation that the camera watches and gives controls back to the player when we start up in a map or come to one the first time.
It currently doesn't work when you leave and come back in the same play.
To do the above, animation was done with the HexTile, Camera watched.
Added in the placeholder sprites to set up more of a world with stuff.
All added to both the hextile script and prefab.
I renamed the HoverAble script to HoverAbleResourceButton, which had to change that reference in the Resource, and ResourceInfoPanel.
In locationManager I took out old Debug logs, and made it do a better check for getting a location for starting point (currently does the bottom left most)
Created a TileInfoPanel script that will be handling all the information that a hextile will hold that a player can utilize.
Currently have a rudimentary version that works upon clicking a Conquered location. Pop up the panel.