Landing/Leaving Sequence Update
Added functionality to the main scene where buttons and calls are made.
In CameraController I changed the name from EndLandingSequence to EndSpaceSequence, as well as the bool inInitializedSequence to inSpaceshipSequence so I can use it both ways.
HexTileInfo got most the work doing the landing in reverse for leaving.
As well as some calls to get it to work whenever it should.
LocationManager now keeps a direct reference of the starter and calls the land/leaving functions when necessary.
Main gained a listener event for when we have decided that we want to leave.
It starts the process by telling the activeBrain to tell it's starter to run the sequence since I didn't want the UI button to know the tile? But it does and could have just accessed it from there... may change that later.
TileInfoPanel looks for whether we are looking at the starter tile and either turns on/off the leave button if true or not.