Resource/Camera/HexTile Update
Resource I changed the non-consumable requirements simply to 1 instead of parsing a non-number.
ResourcePanelInfo now stores a makeable string.
It also temporarily turned off a few things.
TileInfoPanel in awake I removed the listener for the hextileui, decided to go another route. (prob can delete)
HoverableResourceButton removed a blank space.
HexTileInfo added a prefab for a container that can create UI elements.
It also lifts the land tile when the mouse is present.
Moved update to the Unity methods.
Turned off listeners for turning land should an edge piece be asked to turn land.
It creates resource buttons like the "B" function.
Does other functions that deals with the buttons.
Cam controller I made some variables public so that the UI buttons could pay attention to how it is moving.
Changed a lot in the Main Scene to do all the fixes.
Changed the TilePrefab, InfoPanel prefab and planet prefab.
Created a UIMoveWithCamera script that pays attention to the controllers variables and moves it's elements to correspond with the movement.