Tile System Update
Currently commented SaveSystem Location save method for rework.
Removed Debugs in ResourcePanelInfo.
Resource holds a tile which it uses to check for appropriate resources.
Main functions that used Location resources now reverts to Library again.
Main saving and loading method for locations was changed with the new tile info.
LocationManager no longer has it's own resource list.
It also knows how to break down the data and hand it back to tiles when loading and saving.
HoverableResourceButton the bool isHovering was made public to check on the fly.
CameraController sends a message to reposition and size UI, and disable when the zoomed out too much.
UIContainerprefab was altered quite a bit to handle the new ui setup.
HexTileInfo gained its own ResourceData list.
It also holds the camera and the new UIResourceManager script references.
It listens for the camera zoom functions when it has active UI.
It also spawns food resource on all trees currently.
It now builds its own information from the string given of its list of resources when loading.
Prevents UI from activating if the camera is zoomed out.
TileOptionCreation has been given to the new UIResourceManager script.
Has a function to get the tile's troop count to turn on and off the UI Button.
Also updates the resource buttons for checking the hextile resources rather than the ResourceLibrary after it was established.
Added a new script called UIResourceManager to handle all UI needed for a specific tile.