Resource Update
UIcontainer got an update for the troop text when you click on the troop button.
I made some adjustments to the Debug menu so that I could access the information of a given tile and see tile information.
UIResourceManager houses the changes for this.
I put all the Main Que information onto the Hextile script.
Also made a few tweeks to make sure we could see all the resources needed at the start of the game, being food, soldiers and a barracks.
Location manager updates the Starter tile with the correct data for starting on the tile.
Main is what controls the debug information and again, I removed the que data from it.
Resource now has to look to it's tile for the que information rather than main.
UIResourceManager also now rotates all the buttons back to a normal position and does a fun fan out effect based on how many resources are present.