Debug Tool Update
Tile prefab no longer has the UI floating text, but was given to the UIContainer_prefab.
Made a lot of changes in main scene to the debug window.
CameraController looks for the static variable in main to remove UI when zooming out.
HexTile now has functions to reveal all the enemies on the map and to dynamically change the spawn ratio and density.
Location Manager sets all tile enemy values, it is given by main on creation. This will allow us to use Main to change each planets spawn values when they are created.
Main is doing the heavy lifting for all debugging for values. Also gains the can't lose bool so the player doesn't lose any troops.
Main also has a function now that resets the level and passes all the new values that were potentially changed.
UIResourceManager now accesses a slider as a timer when battling.
It also recieves a message from main to destroy itself.
It also needed a new generic container to turn off all the normal tile UI instead of itself.