Debug Update
Created a General script after getting it to work in the Location manager so that it could be duplicated.
Created a prefab for General.
Planetprefab now has a reference to the general prefab.
Lots of changes to the debug menu.
Made changes to the hextile to accommodating a general interacting with it.
LocationManager removed a few unneeded methods, tested the general script in it, it also holds a list of generals.
(will need to set an *active one)
Main got all the UI interactions to call all general methods.
Also updated the Message pop-up.
Set battleTimer to 0 in UIResourceManager when the battle timer needs to start.