Resource Update
Hextileinfo I made it so when all the soldiers from the ship leave, the ship doesn't disappear.
Also changed the name of UIResourceManager in the hextile reference as myUIManager
I got the que to work for a tile, it adds to the que without creating more ques unless the que is another item. It also shows the battle bar as a timer. Also changed the name to be more universal rather than simply battle timer.
If there is an item that was changed from the data sheet, I am having all hextiles run their own comparison for their data.
Main was altered to allow any object to access the comparison method.
Main also had a few things that were old and needed to be removed.
Location manager simply had that reference of the tiles UIResourceManager reference name changed.
I guess I deleted a blank space in resource.