Map Update
Main scene had changes to the debug menu.
Main can now set the viewing level in debug.
It can also set whether you are allowed to see into planets.
(currently only allows viewing of any planet except the first)
Needed a check whether we have physically been to a planetary experience or not.
Needed to adjust a few methods so they each use it.
DepthInteraction now looks to Main's static values whether or not it can work.
Making way for a visual to your planets.
Main can also generate any planet that is consistent with discovery in debug menu.
Location manager now slightly changes a few thigs if we are simply viewing a planet. (doesn't create a starter)
Also no longer stores the data of a planet viewed.
Hextile got a true position for the ship so that it doesn't get messed up if people leave the planet while it is animating.