GeneralsUpdate
Removed the General prefab that I had previously made for the UI. It didn't do what I wanted it to.
I created a whole new one that did do what I wanted.
That being: SetActiveGeneral, show name, location, and status. Activate/Deactivate and choose which tile.
Planet prefab got a change since I was adjusting zooming in and out  with access.
Main scene got a whole GeneralsUI set up.
All GeneralsCode was mode to a new GeneralsContainerManager which creates and manages all generals and their UI.
Hextile got an update so that it wouldn't interact if there are UI over top of them, and will make them go back to normal height if they got stuck.
LocationManager needed a way to communicate with the Generals so it gives itself over to the GeneralsManager.
Main got a few changes since it no longer really needs to do anything with the generals. Also adjusted a method that checks whether we have visited a planet yet or not.
It also contains not universal variables that are checked during various occasions.
Also created a new script that lives on the canvas and watches whether the mouse is over a UI element or not and adjusts a universal check bool.