Ships Update
Added a new script and prefab for shipinventory.
Added a Shipmanager script and a shippanel manager script.
Added a ship script.
Removed the old TileInfoPanel script and UI because they aren't being used any more.
Commented out a debug in the SaveSystem.
Update the UIContainer prefab for the tiles to now have and show different ship options.
Updated the UI to also have a Ship manager UI.
I changed the event in Location manager from onGreetGeneralManager to onGreetManagers.
I use it now for both the UI managers of generals and ships.
I made a change to whether you can look into different levels by allowing full access for the initialization of the first planet.
Camera now gets a message to look at the starter position if it is coming back to a planet already see. So there is no longer the landing sequence when coming back.
TilePrefab got changes so it could handle ship information.
LocationManager now holds a reference to at least one Spacecraft at a time.
It no longer listens for first time interactions on planet start from main, the two methods were combined and more generally used.
It also can give a tile a ship.
Main has quite a bit that allows it to be apart of the ship making and handling process and houses all static values for those.
Each tile now can pay attention to a single ship that may come onto it.