Ships Update
I shrunk the message popup panel because the messages seemed so small in it. Also switched out the button to be the spaceship like the rest.
Made adjustments to the ship info panel template so that it can use a slider to show the amounts of the resources in the tile and the ship and quickly change between them.
It also starts a journey/cancels/and land buttons for the ship to do all those.
UIContainer prefab is where all the tile ui changes took place.
Main scene maintains to take all the changes.
Deptjinteraction I removed a Debug.log
Generals container I renamed the event it was listening to from Location manager and had to fix it.
Tile listens for several instances where one would need the enemyUI to disappear.
It also has several functions where it now can launch and receive a ship.
Also does a check to make sure that it doesn't show a ship when the ship is away.
Location Manager I renamed one of the Actions, and removed one of the GameObject references because it was no longer in use.
When returning to a location I made it so that the camera looks at the original starter tile.
Added in an assignment of the ship when being given a tile.
Also added a new version of the method that should be utilizing a landing sequence or not.
In Main, I added a bool for whether we are landing on a planet or not, so when we access the area we can do that.
It also gained a buildLevelAmount int to ensure that we are only building once and not multiple calls from a button press or however it is getting called.
Changed how the reset works, so that it actually destroys everything and resets all things.
Added a new debug that allows you to set the ship speed.
Removed an unnecessary block of old code.
Moved Application quite to the bottom to be uniform with the other scripts.
Altered how the level is built/handled if we are viewing, landing, or actually needing the level.
Added a GetShipMenu so that we can access it view a tile. (for now)
SaveSystem got a new update to SaveShips, which can be broken down and built back from memory.
ShipsManager can save and load ship data.
Set all ship speeds.
Give ships access to the planet they are supposed to be on.
Assign a tile a ship.
A new CreateAShip method that builds from memory. and Remove all ships.
Shipinventory adjusts the overall cargo and soldier count.
Spacecraft got quite a bit, from UI altering to the actual ship commands.
UIResourceManager pays attention to ship information now and updates UI elements.