UI Update
On the UIContainer for the tiles I switched the resource button call from an old method to CheckResourceButtons.
Main scene was changed for debugging.
Tile got a return ResourceData from string for debugging.
Moved some resource methods to their own area to find them easier.
When a ship leaves a barracks is removed and the buttons will be updated.
When a ship returns/lands a barracks is added to the tile.
Soldier check is now whether we have a ship, any soldier, otherwise turn off UI.
Location Manager returns a tile for Debugging.
Main now holds an active tile for debugging.
Also sets one and even gets a resource off of it.
Spacecraft was given a public string that other scripts can observe - status.
All locations the status text was altered was given a new call to update both the text and the string.
UIResourceManager now holds an active list of the buttons for the tile resources, if the amount on the tile changes the UI will update for the right ones.