Map generation update
The terrain is getting an update that dictates what everything is based on Terrain traits, where some of them are generic right now.
I also added in the three tree and mountain images that we will be using to show the topography of the areas.
Also added a new Item List that will hold all the basic items that can be used to create all things. They will be added to conditions that they meet.