Item System Update
I organized the debug menu and added a call to show all veg, temp, heights, and item distribution.
Added to the topographic name.
Basic Items don't need to be stored since they are apart of the main list.
Therefore I removed all saving of them and just build the list new whether we are loading or getting a new build.
Also put the tile ui into a container in the canvas so it doesn't take up all the space.
Added a setvalue method to the trait script so I can alter them after they have been created.
Also changed the Reset level to Reset all and safereset to reset level.
Made main's isPlanet public and set it when we are building a planet in case it doesn't get set when I normally wanted it to.
Traits also got a library where all are in it, all are save with it, and built from it.
Removed all references to canvas or landFormationOdds since we changed how the algorithm works for generation to perlin noise.
Location manager now builds all tiles, sets all neighbors, sets vegetation levels, then runs the perimeter to make sure they match, then sets all topo graphics, then sets items.
Item also needed a check for whether it's itemData had terrain requirements.